Refactor packages
This commit is a major overhaul of the package structure in this project.
It's the result from the work on the team escalation engine for which
things need to be better separated. This does not yet implement
working team notifications which will be part of a later commit.

Apart from the refactoring this commit also adresses the following issues:
- basic test coverage for (almost) all methods
- go vet clean
- golint clean